refactor: Move some logic for slack activity message formatting into a seprate module.
